Move heading split function to text module
