+++ novelwriter/core/index.py
@@ -43,8 +43,8 @@ from novelwriter.core.indexdata import NOTE_TYPES, TT_NONE, IndexHeading, IndexN
 from novelwriter.core.novelmodel import NovelModel
 from novelwriter.enum import nwComment, nwItemClass, nwItemLayout, nwItemType, nwNovelExtra
 from novelwriter.error import logException
-from novelwriter.text.comments import processComment
 from novelwriter.text.counting import standardCounter
+from novelwriter.text.formats import processComment, processHeading
 
 if TYPE_CHECKING:
     from collections.abc import ItemsView, Iterable
@@ -383,7 +383,7 @@ class Index:
                 continue
 
             if line.startswith("#"):
-                hDepth, hText = self._splitHeading(line)
+                hDepth, hText = processHeading(line)
                 if hDepth == "H0":
                     continue
 
@@ -438,28 +438,10 @@ class Index:
         """Scan an inactive document for meta data."""
         for line in text.splitlines():
             if line.startswith("#"):
-                hDepth, _ = self._splitHeading(line)
+                hDepth, _ = processHeading(line)
                 if hDepth != "H0":
                     nwItem.setMainHeading(hDepth)
                     break
-
-    def _splitHeading(self, line: str) -> tuple[str, str]:
-        """Split a heading into its heading level and text value."""
-        if line.startswith("# "):
-            return "H1", line[2:].strip()
-        elif line.startswith("## "):
-            return "H2", line[3:].strip()
-        elif line.startswith("### "):
-            return "H3", line[4:].strip()
-        elif line.startswith("#### "):
-            return "H4", line[5:].strip()
-        elif line.startswith("#! "):
-            return "H1", line[3:].strip()
-        elif line.startswith("##! "):
-            return "H2", line[4:].strip()
-        elif line.startswith("###! "):
-            return "H3", line[5:].strip()
-        return "H0", ""
 
     def _indexWordCounts(self, tHandle: str, text: str, sTitle: str) -> None:
         """Count text stats and save the counts to the index."""
